Demonstrate Time columns using copier
This commit demonstrates, I think, that we need to ditch copier.

This commit allows us to retrieve a Time field from the database and copy it
into the App Media Type while working within the copier framework.

What I do is create a `{{fieldName}}Time` field and create a
`model.{{fieldName}}()` func which copier will use to translate between the two.
And, yes, I amend the field name from your API def. If you asked for `Time`, I
create `TimeTime`, and you'll have to reference `TimeTime` in your model
accesses.

Another major problem is that coming from the App to the Model, we have no way
to add these helper functions to the source (App) structs.

So, I think we need to come up with something else.